refactor: move reset password logic to client form
